move TimeTravel and PaymentsSummary to right place
